dev: ci

+++ justfile
@@ -4,13 +4,14 @@ default:
 init_system_files:
     #!/bin/zsh
     set -euxo pipefail
-    mkdir -p /home/vscode
-    if [ -f /tmp/.gitconfig ]; then rm -rf /home/vscode/.gitconfig && cp /tmp/.gitconfig /home/vscode/.gitconfig; fi
-    if [ -d /tmp/.ssh ]; then rm -rf /home/vscode/.ssh && cp -r /tmp/.ssh /home/vscode/.ssh && chmod 600 /home/vscode/.ssh/*; fi
-    git config --global --add safe.directory '*'
-    git config --global core.autocrlf input
-    git config --global core.eol lf
-    if [ -f /tmp/.zshrc ]; then cp /tmp/.zshrc /home/vscode/.zshrc; fi
+    if [ "$CI" != true ]; then
+        if [ -f /tmp/.gitconfig ]; then rm -rf /home/vscode/.gitconfig && cp /tmp/.gitconfig /home/vscode/.gitconfig; fi
+        if [ -d /tmp/.ssh ]; then rm -rf /home/vscode/.ssh && cp -r /tmp/.ssh /home/vscode/.ssh && chmod 600 /home/vscode/.ssh/*; fi
+        git config --global --add safe.directory '*'
+        git config --global core.autocrlf input
+        git config --global core.eol lf
+        if [ -f /tmp/.zshrc ]; then cp /tmp/.zshrc /home/vscode/.zshrc; fi
+    fi
 
 
 venv-roots := "/workspace/examples /workspace/plugins/engine_conversions /workspace/plugins/engines /workspace/plugins/runtime_systems"
